PVW-3094: Render CancelledSession
* Add allow dead_code

The CancelledSession is not yet used on the wallet_core side, but will be soon. As such
adding the allow(dead_code) annotation to suppress the warning in the meantime.

* Refactor SessionCancelled -> CancelledSession

Introduced a discrepancy between the core and app by
flipping the order. Now aligning with core by using
CancelledSession everywhere.

* Add test for CoreSessionCancelledError

Verify that the bloc emits a DisclosureSessionCancelledError
when a CoreCancelledSessionError is thrown when trying to accept
the disclosure of requested attributes.

* Refactor SessionStopped -> SessionCancelled

Align with the naming from the wallet core and make
the distinction between session stopped (by user) and
session cancelled (remotely) a bit more distinct.

* Parse & handle CancelledSessionError

* Add trailing commas

* Render DisclosureStopped

Add a DisclosureStopped state and render it in the DisclosureScreen.
Note that this does not yet include the logic to actually figure out
if the session was stopped remotely, hence the //TODO added to the code.
Once that logic is available that part is to be implemented to make the
newly added ui actually show up in this flow.