Revert use of memory-serve crate in MRP
This crate does not function correctly, as it caches the contents of the
assets directory at build time, even in debug mode. When a new hash is
calculated for the main JS file and placed in this directory, re-running
the `mock_relying_party` does not re-read the directory contents and
returns a 404 error when the web-frontend requests this JS file.