Update golden tests
* Update CardDetailErrorScreen

Make it so that the scrollbar doesn't show up
when all content is visible.

* Update CheckData goldens

The padding below the last value was tweaked,
these goldens (that rely on that widget) were yet to be updated.

* Hide unnecessary scrollbars

* Update CheckDataOffering goldens

- Use the ListButton widget for the footer, so that
it has the correct default margins
- Add a no_footer golden to verify correct behaviour of the
last divider in the CheckDataOffering page

* Update CardDetail Loading golden

Set  hasScrollBody: false & fillOverscroll: false to the
SliverFillRemaining widget in the card detail loading state,
so that the content simply takes up the available space and
thus the scrollbar is hidden in the loading state.

* Apply 1 suggestion(s) to 1 file(s)

Co-authored-by: Arjen <edi@minbzk.nl>

* Update PinTimeout Goldens

Make sure the countdown timer shows a consistent value,
by moving the expiry time generation into the widget
so that it's (re)calculated at the last possible time,
causing any delays in the tests to go unnoticed.

* Update Dashboard ErrorState

Add some bottom margin to the dashboard error screen,
so that it doesn't only rely on the OS's safe area.

* Update light confirm action sheet test

* Update several goldens

- Update some generic goldens (flutter upgrade?)
- Update dialog related goldens, so that we only
render a single screensize.

* Update qr_scan_failure

Since this is now a dialog, only generate
1 golden instead of 4 different scenarios.

* Remove my_code_tab

This widget no longer exists and thus doesn't
require a test.

* Add default FlutterAppConfiguration

Add a sane default for the FlutterAppConfiguration that comes out
of the default MockConfigurationRepository, the one fetched through
Mocks.create(). That way the goldens that rely on this don't show a
loader but instead show the provided default values.

* Fix lint issues

* Update golden images

* Update test files

Fix issues with missing dependencies or incorrect surfaces sizes in the
golden tests.