Minor flutter tweaks
* Fix lint issue

* Update Holograph related goldens

Due to the new RepaintBoundary the goldens that included
the holographs were rendered every so slightly different,
so update teh goldens accordingly.

* Provide NavigationService for golden test

* Update NavigationService test

* Add RepaintBoundary

As a slight performance optimization, wrap the
CardHolograph widget inside a RepaintBoundary
widget, this way only the holograph gets rebuild
(based on the gyro output) on every frame instead
of the whole card, resulting in a minor performance
improvement.

* Move processQueue call

Instead of calling the processQueue method through the
DoOnInit widget (which we provide together with the route)
we now add it to the visibility callback that we already use
for semantics announcements. This removes a widget and makes
it so that all this logic is contained within the DashboardScreen
directly.

Also added a sanity assertion to fail early in case the prerequisites
change and are not fulfilled anymore when the user lands on the dashboard,
as we currently assume (by consuming the request) that all prerequisites
are always met when calling processQueue().

* Increase Android scanner resolution

Android builds were having some issues while scanning the QR codes,
increasing the scanner resolution (which only applies to android)
should resolve these issues.