PVW-2832: Support Change Pin (mock only)
* Avoid async gaps

* Disable ChangePin in non-mock

Disable navigation to the ChangePinScreen in non-mock
builds of the app. As the wallet_core will implement the
change pin behaviour later.

* Fix lint/translation issues

* Add ChangePin landscape test

Add a landscape test for the changepin screen so that the landscape
code for the changepinpage is also being hit.

* Fix SetupSecurity Test issue

The previous state was missing the 'error' field in the equals check,
that issue was fixed, causing this test to start failing because a
different error was provided. Fixed by providing the same error in
the expected result.

* Add additional golden tests

Verify all ChangePin states through golden image
tests.

* Add ChangePinScreen widget tests

* Update initial goldens

- Rename files
- Add dark mode initial golden
- Verify title of initial page in widget test

* Add initial golden test

* Verify back and backspace behaviour

* Add happy flow and error bloc tests

* Update ChangePinBloc (Tests)

- Add some documentation and extract methods in the ChangePinBloc.
- Add some bloc tests and utility functions for the upcoming test.

* Lokalise ChangePinScreen

* Extract announceEnteredDigits

Have one common place where we run the pin entry announcements
to avoid duplicating the announcement code.

* Extract leading widget

* Update Lokalise Keys #2

Update the pinValidationErrorDialog lokalise keys so they match
their new location (dialog vs setup security page).

* Update Lokalise Keys

Update the pinConfirmationErrorDialog lokalise keys so they match
their new location (dialog vs setup security page).

* Extract pin dialogs

Extract the validation and confirm error dialogs
so that they can be re-used in the change pin flow.

* Implement fallback for error states

* Update UI for loading/success pages

* Make back button conditional

* Initial mock happy flow

Implement required steps to change the pin for in a mock
build of the app.

* Setup error handling skeleton

* Handle select/confirm error states

* Render repeat pin code UI

* Manage didGoBack

Add and manage the didGoBack flag to show the
correct back animation when the user navigates
back during the change pin flow.

* Introduce FakePagingAnimatedSwitcher

* Configure basics of ChangePinBloc

* Add ChangePinStates

Add the required ChangePinStates to manage the UI

* Reset currentPin when closing PinBloc

Clean up the _currentPin variable when closing the
pin bloc to avoid it lingering in memory longer than
needed.

* Implement initial pin check

* Setup current pin check

Configure the basics to perform the current pin check
in the change pin flow.

* Update debug statement in wallet_core api

* Add ChangePin UI skeleton

* Implement new APIs in core_mock

Implement the change_pin and check_pin apis in the core_mock
implementation so that we can build the change pin UI in the
app mock environment.

* Update wallet_core apis

Setup placeholders for the change_in and the check_pin apis.