PVW-3223: Accessibility Improvements
* Provide locale info

Provide missing locale info to the Text widget in the
select_new_pin_page.dart.

* Align versions

* Fix PVW-3277

The button width calculation broke down because of the new use
of Text.rich widgets throughout the app. This updates the content
width calculation to take the possible occurrence of Text.rich widgets
inside the buttons into account.

* Clean up translations

* Remove stale //FIXME

* Update tests

Fix all the tests that started failing due to the
new use of Text.rich to provide text with locale
hints for VoiceOver/Talkback.

* Fix QR scanner announcements

Add locale attribute.

* Fix lint issue

* Improve announcements

Fix some cases where the announcement did not include locale info.

* Mark Card as Image (CardDetail)

Mark the Card widget as an image on the CardDetailScreen

* Add scan successful announcement

* Move scanning announcement

* Announce QR detected state

* Lokalise language updated announcement

* Announce language change

* Fix minor lint issues

* Migrate to TextSpan

Use TextSpans throughout the app so that we can inform the
VoiceOver/TalkBack about the currently selected language.

* Fix mock data issue

Flipped dutch and english label values

* Rename withActiveLocale to toAttributedString

In preparation of also adding toTextSpan() with similar
functionality but other return type.

* Migrate label to attributedLabel

Migrate from the normal label to the attributedLabel wherever
we specify this in a Semantics widget and also supply the current
locale so that voiceover/talkback pronounces the text correctly.

* Provide extra metadata to ChangeLanguageScreen

 inMutuallyExclusiveGroup: true should be set for Radio like buttons

* Provide locale to Body- & TitleText

This makes sure voiceover/talkback uses the correct TTS
engine to pronounce these texts, even if the device language
differs from language selected in the app.

* Update PinKeyboard focus

Make sure we request focus after the layout so the
user can always use the keyboard to enter the pin
when this widget is shown.

* Fix MockDigiD overlay color

Use orange instead of blue from the theme.

* Add focused color for ElevatedButton

* Add underline to buttons

Add an underline to text in buttons when they are focused
or hovered to give a visual indication of where the
focus currently is.

* Make URLs in OrgDetailScreen focusable

* Fix SharedAttributeCards highlight

When navigating with the 'tab' key this widget did not
show a correct highlight due to mis-placement of the
decoration, this decoration is now pulled up which allows
the material to show the correct touch/selection feedback.

* Update ScrollBar theme

* Fix PinKeyboard focus

Due to re-use of the FocusNode the focus (and thus keyboard
input) did not always work correctly. Now providing a dedicated
focusNode that is tied to the lifetime of the PinKeyboard.

* Exclude DigiD icon semantics

* Update focusColor in theme

* Implement TextWithLink Focus

Introduce the FocusBuilder widget, for an easy way to make items
focus-able and update the UI when they are indeed focused.
Update the TextWithLink to use the new FocusBuilder widget to draw
a grey box around the link part of the text, and trigger the onTap
callback when the user presses the enter key while focused.

* Announce pages in introduction