PVW-2835: Biometrics setup
* Apply 1 suggestion(s) to 1 file(s)

Co-authored-by: Daan <2034049+daandebie@users.noreply.github.com>

* Add convenience method to UnlockMethod enum

* Implement biometrics unlocking in wallet

This implements both saving the unlock method in `Wallet` and unlocking
the wallet without the PIN code, if biometrics are enabled.

* Replace wallet storage update with upsert method

* Update tests

* Refactor UseCase

Instead of calling it supportedBiometrics, now rename it to
availableBiometrics to better reflect what the usecase actually
checks. At a later stage we will likely need a supported usecase
to actually check if the device has biometric support that can be
set up by the user. (in the settings biometrics flow).

* Update SetupSecurityBloc Tests

Take the new authBeforeEnabling flag into account.

* Add mocks and SetBiometrics tests

* Add skipAuth option

Add the option to skip the authentication step
when setting the biometrics flag (useful in later,
settings based flow, where there will be an extra
step after authentication).

* Fix incorrect svg reference

* Add Biometrics Setup goldens

* Add CoreBiometricRepository test

* Set biometric login flag

Pass the enable/disable state of the biometric login
on to the core so that it can be persisted there.

* Update SetupCompleted copy

* Update SetupSecurityCompleted state

Expose whether or not the biometrics were enabled to be able
to show the correct copy on the last page.

* Add & use Face-ID Icon

* Provide localizedReason

Provide a localizedReason when configuring the biometrics.

* Update biometrics core API

Combine the enable/disable api into a single setter.

* Verify setupBiometrics bloc behaviour

- Clean up the pin entry calls
- Verify successful biometrics setup
- Verify skip biometrics setup

* Process biometrics user selection

When the user presses enable, prompt for the biometrics
and enable it accordingly. Or alternatively, when user
opts to skip the biometrics setup simply move to the
completed setup state.

* Add biometrics wallet_core apis

Add basic apis to enable/disable biometrics,
unlock the wallet and get the current state.

* Update SetBiometricsUseCase

Add sanity check before toggling the biometrics
preference.

* Add biometric usecase docs

* Introduce SetBiometricsUseCase

Add a usecase to enable or disable the usage of
biometrics.

* Refine SupportsBiometrics test

Mainly updated docs to make sure test name aligns better
with the tested behaviour.

* Update SupportsBiometrics logic

Log what is supported and tweak the return type
to more accurately match what we need to draw the
correct ui.

* Fix local_auth android 8 issue

App would crash when not using an AppCompat theme
on devices running android <= 8.0.

* Expose SupportedBioMetrics to UI

Use this info to make the title dynamic.

* Return supported biometric types

The UI needs to know which biometrics are supported
to render the correct text. This commit updates the
usecase to expose that info.

* Add placeholder biometrics page

* Update SupportsBiometricsUseCase

Add better iOS support, since it only reports
face of fingerprint type, and not 'strong' as
we expect for android.

* Update SetupSecurityBloc

- Inject the usecase to check for biometrics
- Use the biometrics usecase to emit the correct
state after creating the wallet
- Update the tests to verify this new behavior

* Introduce SupportsBiometricsUseCase

* Add local_auth package

- Add the local_auth package as a dependency
- Add iOS plist config
- Configure android app to support local_auth