PVW-2836: Unlock with Biometrics
* Clean unused strings

* Update semantics copy

* Remove unused file

* Update BiometricsKey label

* Clean up intl files

* Paragraph biometrics text

* Update Switch Design

Match the design as specified in Figma.

* Rename unlock method

* Update test dependencies

* Trigger Biometrics Automatically

Implement logic so that the biometrics are automatically
triggered when the user:
- Opens the app (cold start)
- Returns to the app from background while its locked
(both under the condition that biometrics are enabled).

* Fix BiometricSettings Landscape layout

Make sure safe areas are correctly respected.

* Update LockedOutDialog copy

* Handle LockedOut state

Make sure the user gets some feedback when the biometrics
enters the 'lockedOut' state, i.e. the user provided incorrect
biometrics too many times, causing the system to block biometric
authentication until the user confirms with her pattern/pin when
unlocked the device normally.

* Clean up BiometricSettingsScreen

Add missing spacing parameters.

* Fix PinScreen test

* Add Biometric Unlock support

Add a biometric key to the pinkeyboard and show it on the
PinScreen, when pressed, trigger the biometric unlock flow.

* Make success illustration dynamic

* Update Lokalise strings

* Fix incorrect touchId l10n

* Fix lint issue

* Lokalise BiometricSettingsScreen

* Make settings title dynamic

* Fix iOS enable face-id dialog issue

When showing an unauthorized face when enabling
biometric unlock using face id on iOS the 'setup
biometrics' dialog would be shown due to the current
generic error handling. This has now been updated to
correctly handle the android edge case instead.

Also fixed tests that were broken due to new check biometric
behaviour.

* Add dynamic biometrics title

Expose SupportedBiometrics from the BiometricsSettings
bloc so that the screen can render the correct title.

Migrate from 'isBiometricsSupported' to 'GetSupportedBiometrics'
usecase.

* Biometrics -> Biometric

Refactor classes and rename + document some methods and classes.

* Clean up listen-/buildWhen

* Redirect to Settings

Add the app_settings plugin to redirect users to the app settings
in the edge case described in the comment above its usage (in the
biometric_settings_screen).

* Handle biometric setup required

When enabling the biometrics eagerly (before it is
configured on the device) show the setup required
dialog to guide the user to the correct place.

* Support replacing with TerminalScreen

* Update localizations

* Add confirm setting with pin

Add the UI and logic to confirm the biometric setting
change with pin.

* Introduce TerminalScreen

* Render error state

* Provide bloc and add Builder

* Implement BiometricsBloc basics

* Introduce BiometricsSettingsScreen

* Update Biometrics MenuItem

- Update MenuItem to support subtitle
- Update MenuItem to hide arrow when not clickable
- Add logic to check for biometric support and provide
the correct lokalisations based on this flag.

# Conflicts:
#	wallet_app/lib/src/feature/menu/widget/menu_row.dart

* Introduce CheckDeviceSupportsBiometricsUseCase

* Apply 1 suggestion(s) to 1 file(s)

Co-authored-by: Daan <2034049+daandebie@users.noreply.github.com>

* Add convenience method to UnlockMethod enum

* Implement biometrics unlocking in wallet

This implements both saving the unlock method in `Wallet` and unlocking
the wallet without the PIN code, if biometrics are enabled.

* Replace wallet storage update with upsert method

* Update tests

* Refactor UseCase

Instead of calling it supportedBiometrics, now rename it to
availableBiometrics to better reflect what the usecase actually
checks. At a later stage we will likely need a supported usecase
to actually check if the device has biometric support that can be
set up by the user. (in the settings biometrics flow).

* Update SetupSecurityBloc Tests

Take the new authBeforeEnabling flag into account.

* Add mocks and SetBiometrics tests

* Add skipAuth option

Add the option to skip the authentication step
when setting the biometrics flag (useful in later,
settings based flow, where there will be an extra
step after authentication).

* Fix incorrect svg reference

* Add Biometrics Setup goldens

* Add CoreBiometricRepository test

* Set biometric login flag

Pass the enable/disable state of the biometric login
on to the core so that it can be persisted there.

* Update SetupCompleted copy

* Update SetupSecurityCompleted state

Expose whether or not the biometrics were enabled to be able
to show the correct copy on the last page.

* Add & use Face-ID Icon

* Provide localizedReason

Provide a localizedReason when configuring the biometrics.

* Update biometrics core API

Combine the enable/disable api into a single setter.

* Verify setupBiometrics bloc behaviour

- Clean up the pin entry calls
- Verify successful biometrics setup
- Verify skip biometrics setup

* Process biometrics user selection

When the user presses enable, prompt for the biometrics
and enable it accordingly. Or alternatively, when user
opts to skip the biometrics setup simply move to the
completed setup state.

* Add biometrics wallet_core apis

Add basic apis to enable/disable biometrics,
unlock the wallet and get the current state.

* Update SetBiometricsUseCase

Add sanity check before toggling the biometrics
preference.

* Add biometric usecase docs

* Introduce SetBiometricsUseCase

Add a usecase to enable or disable the usage of
biometrics.

* Refine SupportsBiometrics test

Mainly updated docs to make sure test name aligns better
with the tested behaviour.

* Update SupportsBiometrics logic

Log what is supported and tweak the return type
to more accurately match what we need to draw the
correct ui.

* Fix local_auth android 8 issue

App would crash when not using an AppCompat theme
on devices running android <= 8.0.

* Expose SupportedBioMetrics to UI

Use this info to make the title dynamic.

* Return supported biometric types

The UI needs to know which biometrics are supported
to render the correct text. This commit updates the
usecase to expose that info.

* Add placeholder biometrics page

* Update SupportsBiometricsUseCase

Add better iOS support, since it only reports
face of fingerprint type, and not 'strong' as
we expect for android.

* Update SetupSecurityBloc

- Inject the usecase to check for biometrics
- Use the biometrics usecase to emit the correct
state after creating the wallet
- Update the tests to verify this new behavior

* Introduce SupportsBiometricsUseCase

* Add local_auth package

- Add the local_auth package as a dependency
- Add iOS plist config
- Configure android app to support local_auth