PVW-3226: Updated StepperIndicator
* Update goldens related to StepperIndicator

Also fix a missing provider for the SettingsScreen to make sure
that the test runs successfully.

* Improve StepperIndicator

- Split out single cell into widget (performance)
- Build the main portion of the progress
bar in dedicated method (readability)

* Add sanity check for stepper length

* Add basic stepper animation

Animate the StepperIndicator by growing/shrinking the main progress
indicator and fading in/out the indicators of how many steps are left.

* Redesign StepperIndicator

Basic implementation of the new stepper indicator.