PVW-3454 Root/Jailbreak detection
* Update test

- Re-add (cut/paste mistake) assertion
- Add some more docs

* Simplify RootCheck behaviour

Remove usage of the `afterFirstLayout` callback, as the
root detection can be performed even before the first layout,
simply in the initState.

* Update RootDetectedScreen copy

* Add RootDetectedScreen test

* Import copy from Lokalise

Add the following keys and use them:
- rootDetectedScreenTitle
- rootDetectedScreenDescription

* Introduce RootDetectedScreen

Set up a widget that is to be shown when root is detected
on the device. Dependencies should be kept to a minimum so
that we only need to provide the bare minimum. (we do this
because we want to avoid leaking any data through provided
dependencies, this way the RootChecker can be placed at the
top of the widget tree, meaning no dependencies are provided
or instantiated yet).

* Add extra RootChecker test

* Update RootChecker placeholder

* Add RootChecker test

* Implement and use RootChecker

Some changes to the PreferredLocaleCubit where needed as with the
new RootChecker widget replacing the other tree, this class started
throwing because it emitted after the cubit was closed.

* Introduce RootChecker Widget

Add an entry to perform the root/jailbreak check.