PVW-3380: Fix scanner orientation bug
* Explicitly manage cameraController

The cameraController was managed by the MobileScanner
widget (through the autostart flag). But (not sure if its
due to an update?) it does not seem to be fully disposed
internally. As such, opting to fully manage the controller
in the parent widget now, i.e. disposing it in the onDispose
callback.

* Document quarterTurns

* Split out method to resolve turns

* Fix scanner orientation bug

The scanner is always initialized in portrait, causing the
viewfinder feed to be rotated incorrectly when the device is
in a non-portrait orientation. We now observe this (native)
rotation and rotate the viewfinder accordingly. We rely on a
package here because the flutter built-in orientation does not
provide enough info (no left-/rightLandscape). The inner widgets
are rotated back again so that all the ui (except for the
viewfinder) is still oriented as expected.