PVW-3515: Add RP name to policy texts
* Move OrganizationPolicy to file

* Refactor PolicyContainer

Turn it into OrganizationPolicy to make it a bit
more explicit what it contains.

* Refresh goldens

The branch with the new outline was merged, updating the goldens
that contain the new agreement section AND shared attributes card.

* Fix lint issue

* Fix policy_screen_test.dart

* Update PolicyScreen

Add the organization name in the last section.

* Fix disclosure_confirm_data_attributes_page_test

* Update disclosure_screen_test

Make sure the agreement section is also verified
in the golden test.

* Update PolicyMapper

Update the policy mapper so that it also takes in the
organization which is now rendered together with the
terms of the policy.