Move AppleAttestedKey to wallet_common, add mock
As `wallet_common` will need the `AppleAttestedKey` trait and its
associated `AppleAssertion` for verifying signed messages sent from the
Wallet to the Wallet Provider, this moves them to this crate.

It also implements a `MockAppleAttestedKey` type that implements this
trait and uses the `mock` feature on the `apple_app_attest` crate to
create mock assertions. This is to be used in tests.