Introduce LocalEcdsaKey type for testing
This type performs the role of `RemoteEcdsaKey` during testing and is
the key type produced by `LocalKeyFactory`. It replaces the use of
`SoftwareEcdsaKey` in many tests, which prevents the latter type
fulfilling two roles, i.e. as a substitute for `HardwareEcdsaKey` and as
a substitute for `RemoteEcdsaKey`.

This also includes a number of crate feature dependency cleanups.