Add SelfSignedCa type to replace KeyPair for CAs
The `SelfSignedCa` type can be used when generating CAs for both testing
and in the `wallet_ca` utility. This replaces the use of `KeyPair` for
CAs, the internal representation of which is only suitable for leaf
certificates, not CAs.

This also includes some refactoring of test code and the `wallet_ca`
utility.