Add Android keys to MockHardwareAttestedKeyHolder
This refactors `MockHardwareAttestedKeyHolder` to add support for
producing mock Android attested keys, in addition to mock Apple attested
keys. It also changes this type so that, except for
`PersistentMockAttestedKeyHolder`, the key state is no longer global to
the application, but restricted to the type.

This also updates a number of tests to take the refactor changes into
account.