PVW-3929: Clean up high-severity Sonarqube issues
* Run tests on non-mock builds

Previously the test ran with the '--dart-define=MOCK_REPOSITORIES=true' flag, causing it
to test some mock-flows. This commit removes that flag so the tests are run with the default
non-mock implementation. A few commits ago (on this branch) the test code has been updated
to still verify the mock QR code behavior as intended.

* Clean up old comment

* Clean up yaml files

* Lint fixes

* Fix CoreWalletRepository test

Check for the new (more appropriate) StateError.

* Fix QrCodeRepository tests

* Clean up mock_organizations.dart

* Clean up mock_issuance_responses.dart

* Clean up mock_disclosure_requests.dart

* Clean up other_styles_tab.dart

* Reduce pin_page complexity

Break up the build method

* Make assertions more explicit

* Define constant not-registered error

* Update l10n.yaml conventions

* Fix analysis_options.yaml style issue