PVW-3929: Fix Sonarqube medium severity issues
* Remove stale FIXME

The fixme described here is insignificant and most likely
doesn't impact performance in any meaningful way. As such
removing the //FIXME identifiers to drop it's severity
level. Keeping the note so that, if in the future this
code is touched, we now what behaviour to validate.

* Remove invalid //FIXME

Flagged by sonarqube, but this doesn't have to be fixed,
it's mostly a note about something that could be optimized.

* Simplify ErrorScreen.show()

Reduce the number of parameters used in the
show() method by combining all UI related
parameters into a single wrapper class.

* Simplify TerminalScreen.show()

Reduce the number of parameters used in the
show() method by combining all UI related
parameters into a single wrapper class.

* Fix .yaml warnings