PVW-3809 Update PID flow
* Revert imported Lokalise changes

Revert some changes as they are not part of this feature
and need to be reconsidered.

* Refactor to MissingAttributeRow

Keep the class naming more inline with other attribute
widgets. Also remove no longer valid (image attribute)
related test.

* Format generated files

* Format mocks

* Add appStore translations

Make sure they are synced from Lokalise with the script as well,
by adding the file and 'other' platform in the lokalise web ui.

* Update translations (incl. clean up)

* Remove obsolete .map

We no longer need to map the previewAttributes to
UiAttributes to make them render correctly during
the onboarding flow, so removing the mapper (which
was inconveniently placed in the UI layer to begin
with anyway).

* Update Attribute docs

* Update CheckDataPage goldens

* Update Attribute widgets

* Update AttributeRow

Update the AttributeRow to use the sealed class
properly.

* Turn Attribute into Sealed class