Improve Fastfile fvm or flutter detection
Using a naked rescue is bad practice and also the command shows as blown up.
Remove the disable analytics lane as this is already done in the CI images.
Remove the FASTLANE_OPT_OUT_USAGE environment variable as opt_out_usage is set in Fastfile.
Use `UI.message` instead of `puts` as recommended.