Add mock IntegrityVerdict feature to WP
This feature should only be enabled for local development when using
the Android emulator and/or when the developer does not have access to
the Google Cloud service account required to request integrity verdicts.

This also introduces a convenience feature `android_emulator`, which
combines this feature and `allow_android_emulator_keys`.