PVW-4034: Update wallet_app error handling
* Remove upperCase RedirectErrors

The uppercase variants were added as a fallback until the core issue was fixed,
since there is now an open MR to fix this issue, that should be resolved soon.
As such, removing unnecessary extra keys here.

* Introduce handleAppError to handle stream errors

Introduce an extension method to consistently handle how streams
in usecases catch and handle errors.

* Clean up basic usecases

Many usecases had a very similar behaviour, where they would expose the
results of a method call while handling the error. All those usecases
are simplified in this commit by routing that method call through a
standardized tryCatch method.

* Setup WalletUseCase and extend all UseCases

* Rename GenericError field

Make it more explicit that the value is not intended for the end user.

* Make ApplicationError.sourceError required

Update the ApplicationError class to make the error field mandatory.

* Update provider doc

Fix stale reference

* Revert incorrect ErrorState implementation

Events should not implement ErrorState, only State objects should do this.

* Update wallet_personalize_bloc

Emit states directly when the emitter is available (instead of routing it through
the UpdateState event).

* Update WalletPersonalizeEvent to implement ErrorState

* Rename files

* Migrate CheckHasInternetUseCase to NetworkRepository

Instead of relying on a usecase directly to check the internet status,
move this logic into a NetworkRepository class. The intention is to eventually
inject this into a Mapper<CoreError, ApplicationError> class that can be injected
in usecases. This way we don't have to rely on a static instance anymore in the
future.

* Update ResetWalletUseCase

Update ResetWalletUseCase to throw as StateError when it fails,
since failing to reset the wallet puts the app in an unexpected
state we will throw a StateError and exit the app.

* Remove unused usecases

Get rid of two legacy, unused usecases.

* Update ObserveWalletLockedUseCase

Update ObserveWalletLockedUseCase to make sure it only exposes
ApplicationErrors in case of exceptions.

* Update GetVersionStringUseCase

Update GetVersionStringUseCase to return a Result<> type response and
use this accordingly in the relevant tests/services.

* Update DecodeUriUseCase

Update DecodeUriUseCase to return a Result<> type response and
use this accordingly in the relevant tests/services.

* Update ObserveVersionStateUsecase

Make sure it only exposes ApplicationErrors and not simply passes on
any incoming exception.

* Update StartSignUseCase

Update the StartSignUseCase to use the Result<> return type and
update related blocs accordingly.

* Update RejectSignAgreementUseCase

Update RejectSignAgreementUseCase to use the Result<> return type,
also update the bloc to rely on this new form.

* Update DecodeQrUseCase

Update DecodeQrUseCase to use the Result<> return type,
also update related usecases and tests to rely on this.

* Fix CheckIsValidPin related tests

* Update CheckIsValidPinUseCase

Update CheckIsValidPinUseCase to use the new Result<> return type,
also update the related blocs.

* Update ChangePinUseCase related tests

* Update ChangePinUseCase

Update the ChangePinUseCase to use the Result<> return type,
and update the change_pin_bloc to handle these changes.

* Fix WalletPersonalizeBloc related tests

Remove leftover try{}catch{} clauses from wallet_personalize_bloc
so it relies on the new Result<> format.

* Update rest of WalletPersonalizeBloc

Remove leftover try{}catch{} clauses from wallet_personalize_bloc
so it relies on the new Result<> format.

* Update ContinuePidIssuanceUseCase

Update ContinuePidIssuanceUseCase to use the Result<> return type,
and update teh wallet_personalize_bloc accordingly to adapt to these
changes.

* Update CancelPidIssuanceUseCase

Use the Result<> return type.

* Update StartIssuanceUseCase

Use the Result<> return type.

* Update ContinueIssuanceUseCase

Use the Result<> return type.

* Update CancelIssuanceUseCase

Use the Result<> return type.

* Update AcceptIssuanceUseCase

Return Result<> from AcceptIssuanceUseCase

* Update ObserveRecentHistoryUsecase

Make sure it only exposes ApplicationErrors

* Streamline CoreError->ApplicationError

Create an error_extension.dart file so there is a single
and consistent way to map errors to ApplicationErrors.
Usecases can still opt to handle errors differently based
on their requirements but mostly, catching the CoreError and
mapping it with this class should do the trick and result in
consistent error handling.

* Update ObserveRecentWalletEventsUseCase

Update the usecase so it only exposes ApplicationErrors.

* Update GetWalletEventsUsecase

Update the usecase to follow the new Result<> pattern.

* Introduce start_disclosure_usecase_impl_test

Validate that the start_disclosure_usecase exposes the errors as
expected (i.e. how they were previously handled and tested in the
disclosure_bloc). Since this could no longer be validated in the disclosure_bloc_test itself.

* Update DisclosureBloc and relatives

Update the usecases used by the DisclosureBloc to conform to the new
Result<> response mechanism. Making sure the DisclosureBloc still
behaves as expected (including the tests) required a lot of changes
throughout multiple classes (e.g. propagating the ApplicationErrors
properly from the PinBloc). As such this resulted in many changes
in this one commit. The result however, is a (still) functioning
DisclosureBloc (and app).

* Update CheckPinUseCase (and related usecases)

Update them to use the new Result<> return type and update
the PinBloc and related classes to handle this accordingly.

* Update ObserveWalletCard(s)UseCase

Make sure it emits an ApplicationError when the stream throws.

* Update ObserveWalletCardDetailUseCaseImpl

Make sure it emits an ApplicationError when the stream throws.

* Update LockWalletUseCase usecase

Also move it to a more appropriate location.

* Update GetWalletCards usecase

* Update GetWalletCard And GetWalletEvents usecases

Additionally add some helper methods to the Result class to make working with multiple Result responses a bit more convenient.

* Guarantee Biometric availability checks won't throw

Add a catch-all clause returning a sane value if checking for
biometric availability fails.

* Update BiometricLoginEnabledUseCase

Throw an explicit error to indicate that this check should never fail.

* Update Biometric related usecases

Make it return a Result<> and handle that accordingly.

* Update SetBiometricsUseCase

Make it return a Result<> and handle that accordingly.

* Update CreateWalletUseCase

Make it return a Result<> and handle that accordingly.

* Update init related tests

* Update init/checkPid usecases and related BLoCs

Mainly remove the try/catch inside of the BLoCs, as the usecase
should now guarantee they do not throw any exception, as otherwise
they should implement Result<> as a return type.

Note: in this case the usecases do actually throw, but these are
StateErrors which should never be handled gracefully, they imply
and incorrect state of the app and should thus result in a crash.

* Introduce Error/Result class

Introduce new error and result classes, with the aim of having
usecases always return errors through this new result class mechanism.
This way the usecases will no longer be responsible for try/catch-ing
the usecase.invoke() methods, as it's the responsibility of the usecase
class to explicitly define whether or not a usecase can return an error.
If it simply returns a type directly (without Result<>) it means the
usecase will not throw.